moved xgo to be used on all routes and internalServerMiddleware
