+++ routes/person.go
@@ -5,7 +5,6 @@ import (
 	"github.com/stakwork/sphinx-tribes/auth"
 	"github.com/stakwork/sphinx-tribes/db"
 	"github.com/stakwork/sphinx-tribes/handlers"
-	"github.com/stakwork/sphinx-tribes/utils"
 )
 
 func PersonRoutes() chi.Router {
@@ -14,7 +13,7 @@ func PersonRoutes() chi.Router {
 	r.Group(func(r chi.Router) {
 		r.Get("/{pubkey}", peopleHandler.GetPersonByPubkey)
 		r.Get("/id/{id}", peopleHandler.GetPersonById)
-		r.Get("/uuid/{uuid}", utils.AutoLog(peopleHandler.GetPersonByUuid))
+		r.Get("/uuid/{uuid}", peopleHandler.GetPersonByUuid)
 		r.Get("/uuid/{uuid}/assets", handlers.GetPersonAssetsByUuid)
 		r.Get("/githubname/{github}", handlers.GetPersonByGithubName)
 	})
